add logic to serve new client
- for simplicity, the program will make new allocation of dynamic memory for each new client.

so, it just pre-allocate the pointer (which 48-bit size in most 64-bit modern computer) instead of pre-allocate the struct

it is still possible to choose 'pre-allocate the struct' way, but it may need extra consideration, and still have tradeoff when compared to pre-allocated pointer, e.g from size perspective, the pre-allocate pointer may more optimal, but from performance perspective, you don't need to allocate each time the program serving new client, and thus pre-allocate the struct may more optimal.

but either way, to make it more optimal, I still need to consider how to re-use/lookup for an empty slot to prevent auto allocation when possible (this optimization may be a top priority on constrained environment, where memory space is limited and the computer is very fast to perform such lookup)

for now, the program utilize epoll mechanism to pick the right pointer to the struct, so we don't need to create an algorithm to pick the right item with index.

add function to convert network-byte address into a readable, formatted string.